+++ pilot/pkg/networking/core/v1alpha3/listener.go
@@ -92,18 +92,6 @@ const (
 	// virtualInboundCatchAllHTTPFilterChainName is the name of the catch all http filter chain
 	virtualInboundCatchAllHTTPFilterChainName = "virtualInbound-catchall-http"
 
-	// WildcardAddress binds to all IP addresses
-	WildcardAddress = "0.0.0.0"
-
-	// WildcardIPv6Address binds to all IPv6 addresses
-	WildcardIPv6Address = "::"
-
-	// LocalhostAddress for local binding
-	LocalhostAddress = "127.0.0.1"
-
-	// LocalhostIPv6Address for local binding
-	LocalhostIPv6Address = "::1"
-
 	// ProxyInboundListenPort is the port on which all inbound traffic to the pod/vm will be captured to
 	// TODO: allow configuration through mesh config
 	ProxyInboundListenPort = 15006
@@ -167,7 +155,7 @@ func (configgen *ConfigGeneratorImpl) buildSidecarInboundListeners(
 
 	sidecarScope := node.SidecarScope
 	noneMode := node.GetInterceptionMode() == model.InterceptionNone
-	// No user supplied sidecar scope or the user supplied one has no ingress listeners
+	// No user supplied sidecar scope or the user supplied one has no ingress listeners.
 	if !sidecarScope.HasIngressListener() {
 		// Construct inbound listeners in the usual way by looking at the ports of the service instances
 		// attached to the proxy
@@ -196,7 +184,7 @@ func (configgen *ConfigGeneratorImpl) buildSidecarInboundListeners(
 		//              Address:Port 172.16.0.1:3333
 		//              ServicePort 9999|Unknown
 		//
-		//	The pilot will generate three listeners, the last one will use protocol sniffing.
+		//	Pilot will generate three listeners, the last one will use protocol sniffing.
 		//
 		for _, instance := range node.ServiceInstances {
 			endpoint := instance.Endpoint
@@ -207,13 +195,10 @@ func (configgen *ConfigGeneratorImpl) buildSidecarInboundListeners(
 			wildcard, _ := getActualWildcardAndLocalHost(node)
 			bind := wildcard
 
-			// Local service instances can be accessed through one of three
-			// addresses: localhost, endpoint IP, and service
-			// VIP. Localhost bypasses the proxy and doesn't need any TCP
-			// route config. Endpoint IP is handled below and Service IP is handled
-			// by outbound routes.
-			// Traffic sent to our service VIP is redirected by remote
-			// services' kubeproxy to our specific endpoint IP.
+			// Local service instances can be accessed through one of three addresses: localhost, endpoint IP,
+			// and service VIP. Localhost bypasses the proxy and doesn't need any TCP route config. Endpoint IP
+			// is handled below and Service IP is handled by outbound routes. Traffic sent to our service VIP is
+			// redirected by remote services' kubeproxy to our specific endpoint IP.
 			port := *instance.ServicePort
 			port.Port = int(endpoint.EndpointPort)
 			listenerOpts := buildListenerOpts{
@@ -236,7 +221,6 @@ func (configgen *ConfigGeneratorImpl) buildSidecarInboundListeners(
 			}
 		}
 		return listeners
-
 	}
 
 	for _, ingressListener := range sidecarScope.Sidecar.Ingress {
@@ -1544,34 +1528,6 @@ func (ml *MutableListener) build(opts buildListenerOpts) error {
 	}
 
 	return nil
-}
-
-// getActualWildcardAndLocalHost will return corresponding Wildcard and LocalHost
-// depending on value of proxy's IPAddresses.
-func getActualWildcardAndLocalHost(node *model.Proxy) (string, string) {
-	if node.SupportsIPv4() {
-		return WildcardAddress, LocalhostAddress
-	}
-	return WildcardIPv6Address, LocalhostIPv6Address
-}
-
-func getPassthroughBindIP(node *model.Proxy) string {
-	if node.SupportsIPv4() {
-		return util.InboundPassthroughBindIpv4
-	}
-	return util.InboundPassthroughBindIpv6
-}
-
-// getSidecarInboundBindIP returns the IP that the proxy can bind to along with the sidecar specified port.
-// It looks for an unicast address, if none found, then the default wildcard address is used.
-// This will make the inbound listener bind to instance_ip:port instead of 0.0.0.0:port where applicable.
-func getSidecarInboundBindIP(node *model.Proxy) string {
-	// Return the IP if its a global unicast address.
-	if len(node.GlobalUnicastIP) > 0 {
-		return node.GlobalUnicastIP
-	}
-	defaultInboundIP, _ := getActualWildcardAndLocalHost(node)
-	return defaultInboundIP
 }
 
 func mergeTCPFilterChains(incoming []*listener.FilterChain, listenerOpts buildListenerOpts, listenerMapKey string,
